Lib: move all free radicals to toolbox and update API usage.
