+++ applications/ibutton/ibutton-cli.cpp
@@ -2,7 +2,7 @@
 #include <api-hal.h>
 #include <stdarg.h>
 #include <cli/cli.h>
-#include <args.h>
+#include <lib/toolbox/args.h>
 
 #include "helpers/key-info.h"
 #include "helpers/key-worker.h"
