+++ lib/app-scened-template/file-worker.c
@@ -1,6 +1,6 @@
 #include "file-worker.h"
 #include "m-string.h"
-#include <hex.h>
+#include <lib/toolbox/hex.h>
 #include <dialogs/dialogs.h>
 #include <furi.h>
