+++ lib/toolbox/hex.c
