+++ lib/toolbox/hex.h
